% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_query_text}
\alias{clean_query_text}
\title{Clean text for URL encoding}
\usage{
clean_query_text(text)
}
\arguments{
\item{text}{Text string to clean}
}
\value{
URL-encoded string
}
\description{
Clean text for URL encoding
}
\keyword{internal}
