% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_results}
\alias{flatten_results}
\title{Convert nested lists to data frame}
\usage{
flatten_results(results)
}
\arguments{
\item{results}{List of API results}
}
\value{
data.frame
}
\description{
Convert nested lists to data frame
}
\keyword{internal}
