% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_response}
\alias{parse_response}
\title{Parse API response}
\usage{
parse_response(response)
}
\arguments{
\item{response}{Raw API response}
}
\value{
Parsed response object
}
\description{
Parse API response
}
\keyword{internal}
