% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_params}
\alias{validate_params}
\title{Validate query parameters}
\usage{
validate_params(params)
}
\arguments{
\item{params}{List of query parameters}
}
\value{
Logical indicating if parameters are valid
}
\description{
Validate query parameters
}
\keyword{internal}
